# backend/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  gorm.Model:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Answer:
    properties:
      answer:
        type: string
      correct:
        type: boolean
      model:
        $ref: '#/definitions/gorm.Model'
      question_id:
        type: integer
    type: object
  models.ApiError:
    properties:
      message:
        type: string
    type: object
  models.ApiMessage:
    properties:
      message:
        type: string
    type: object
  models.Location:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.Answer'
        type: array
      model:
        $ref: '#/definitions/gorm.Model'
      question:
        type: string
      survey_id:
        type: integer
      weight:
        type: integer
    type: object
  models.Sector:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.Skill:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.Survey:
    properties:
      obtention_rate:
        type: integer
      questions:
        items:
          $ref: '#/definitions/models.Question'
        type: array
    type: object
  models.User:
    properties:
      age:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.Location'
        type: array
      model:
        $ref: '#/definitions/gorm.Model'
      role:
        type: integer
      sectors:
        items:
          $ref: '#/definitions/models.Sector'
        type: array
      skills:
        items:
          $ref: '#/definitions/models.Skill'
        type: array
      survey_score:
        description: Percentage, out of a hundred
        type: integer
      videos:
        items:
          $ref: '#/definitions/models.Video'
        type: array
      views:
        type: integer
    type: object
  models.UserLocationInput:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
  models.UserSectorInput:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
  models.UserSkillInput:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
  models.UserUpdateForm:
    properties:
      age:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.UserLocationInput'
        type: array
      role:
        type: integer
      sectors:
        items:
          $ref: '#/definitions/models.UserSectorInput'
        type: array
      skills:
        items:
          $ref: '#/definitions/models.UserSkillInput'
        type: array
      videos:
        items:
          $ref: '#/definitions/models.UserVideoInput'
        type: array
    type: object
  models.UserVideoInput:
    properties:
      content:
        type: string
      id:
        type: integer
    type: object
  models.Video:
    properties:
      link:
        type: string
      model:
        $ref: '#/definitions/gorm.Model'
      user_id:
        type: integer
    type: object
  routes.healthResponse:
    properties:
      db_status:
        type: string
      version:
        type: string
    type: object
  routes.loginBody:
    properties:
      mail:
        type: string
      password:
        type: string
    required:
    - mail
    - password
    type: object
  routes.loginResponse:
    properties:
      token:
        type: string
    type: object
  routes.registerBody:
    properties:
      age:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      mail:
        type: string
      password:
        type: string
      role:
        type: integer
    required:
    - age
    - first_name
    - last_name
    - mail
    - password
    - role
    type: object
  routes.userSubmitAnswer:
    properties:
      checked:
        type: boolean
      id:
        type: integer
    required:
    - checked
    - id
    type: object
  routes.userSubmitBody:
    properties:
      questions:
        items:
          $ref: '#/definitions/routes.userSubmitQuestion'
        type: array
    required:
    - questions
    type: object
  routes.userSubmitQuestion:
    properties:
      answers:
        items:
          $ref: '#/definitions/routes.userSubmitAnswer'
        type: array
      id:
        type: integer
    required:
    - answers
    - id
    type: object
  routes.userSubmitResponse:
    properties:
      percentage:
        type: integer
    type: object
  routes.videoLinkUploadBody:
    properties:
      link:
        type: string
    required:
    - link
    type: object
  routes.videoPaginatedResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
      video:
        $ref: '#/definitions/models.Video'
    type: object
info:
  contact: {}
paths:
  /account/login:
    post:
      consumes:
      - application/json
      description: Account login endpoint
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.loginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Account login endpoint
      tags:
      - Account
  /account/register:
    post:
      consumes:
      - application/json
      description: Account register endpoint
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.registerBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Account register endpoint
      tags:
      - Account
  /health:
    get:
      consumes:
      - application/json
      description: Health check
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.healthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/routes.healthResponse'
      summary: Health check
      tags:
      - Utils
  /ping:
    get:
      consumes:
      - application/json
      description: Simply says pong
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: ping
      tags:
      - Utils
  /ping/auth:
    get:
      consumes:
      - application/json
      description: Simply says pong
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: ping
      tags:
      - Utils
  /survey:
    get:
      consumes:
      - application/json
      description: Get the full survey
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Survey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get the full survey
      tags:
      - Survey
    post:
      consumes:
      - application/json
      description: Initialize the first survey
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Survey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Survey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Initialize the first survey
      tags:
      - Survey
    put:
      consumes:
      - application/json
      description: Modify the survey
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Survey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Survey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Modify the survey
      tags:
      - Survey
  /survey/submit:
    post:
      consumes:
      - application/json
      description: User submitted answers to the survey, used to calculate the score
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.userSubmitBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.userSubmitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: User submitted answers to the survey
      tags:
      - Survey
  /users:
    get:
      consumes:
      - application/json
      description: Get all users (paginated)
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all users (paginated)
      tags:
      - Users
  /users/:id:
    delete:
      consumes:
      - application/json
      description: Delete a singular user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete a singular user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a singular user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get a singular user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a singular user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateForm'
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update a singular user
      tags:
      - Users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the current user's details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get the current user's details
      tags:
      - Users
  /videos:
    get:
      consumes:
      - application/json
      description: Get a feed of videos
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.videoPaginatedResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get a feed of videos
      tags:
      - Videos
  /videos/link:
    post:
      consumes:
      - application/json
      description: Upload a video link, must be direct links to files (.mp4)
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.videoLinkUploadBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Video'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Upload a video link
      tags:
      - Videos
swagger: "2.0"